* commit '120ec06344ffe0f282d8dd7ee9c65f51da683f10':
  适配：通过VMware页面删除vmfs，重新挂载vmfs列表不展示问题
  调用卷详情接口改用多线程
  调用卷详情接口改用多线程

+++ dmestore-service/src/main/java/com/huawei/dmestore/services/VmfsAccessServiceImpl.java
@@ -49,6 +49,9 @@ import java.util.List;
 import java.util.Map;
 
 import java.util.Set;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 import jdk.nashorn.internal.parser.Parser;
 import sun.rmi.runtime.Log;
@@ -201,8 +204,12 @@ public class VmfsAccessServiceImpl implements VmfsAccessService {
             // 整理数据
             Map<String, String> stoNameMap = getStorNameMap(storagemap);
 
+            Map<String, VmfsDataInfo> volIds = new HashMap<>();
+
             // 取得vcenter中的所有vmfs存储。
+            long start = System.currentTimeMillis();
             String listStr = vcsdkUtils.getAllVmfsDataStoreInfos(DmeConstants.STORE_TYPE_VMFS);
+            LOG.info("取得vcenter中的所有vmfs存储时间：{}ms", System.currentTimeMillis() - start);
             if (!StringUtils.isEmpty(listStr)) {
                 JsonArray jsonArray = new JsonParser().parse(listStr).getAsJsonArray();
                 if (jsonArray != null && jsonArray.size() > 0) {
@@ -224,9 +231,26 @@ public class VmfsAccessServiceImpl implements VmfsAccessService {
                             vmfsDataInfo.setObjectid(ToolUtils.jsonToStr(jo.get(OBJECTID)));
 
                             DmeVmwareRelation dvr = dvrMap.get(vmwareStoreobjectid);
-                            getVmfsDetailFromDme(relists, stoNameMap, vmfsDataInfo, dvr.getVolumeId());
+                            volIds.put(dvr.getVolumeId(), vmfsDataInfo);
                         }
                     }
+                    Iterator<String> iterator = volIds.keySet().iterator();
+                    int k = 0;
+                    Map<String, VmfsDataInfo> tm = new HashMap<>();
+                    long start1 = System.currentTimeMillis();
+                    while(iterator.hasNext()){
+                        k++;
+                        String key = iterator.next();
+                        tm.put(key, volIds.get(key));
+                        if (k % 20 == 0){
+                            getVmfsSync(tm, relists, stoNameMap);
+                            tm = new HashMap<>();
+                        }
+                    }
+                    if (k % 20 > 0){
+                        getVmfsSync(tm, relists, stoNameMap);
+                    }
+                    LOG.info("调用vmfs存储接口时间：{}ms", System.currentTimeMillis() - start1);
                 }
             } else {
                 LOG.info("list vmfs return empty");
@@ -236,6 +260,22 @@ public class VmfsAccessServiceImpl implements VmfsAccessService {
             throw new DmeException(e.getMessage());
         }
         return relists;
+    }
+
+    public synchronized void getVmfsSync(Map<String, VmfsDataInfo> volIds, List<VmfsDataInfo> vmfsDataInfos, Map<String, String> stoNameMap){
+        ExecutorService executorService = Executors.newFixedThreadPool(volIds.size());
+        CountDownLatch countDownLatch = new CountDownLatch(volIds.size());
+        for (Map.Entry<String, VmfsDataInfo> entry: volIds.entrySet()){
+            executorService.execute(()->{
+                getVmfsDetailFromDme(vmfsDataInfos, stoNameMap, entry.getValue(), entry.getKey());
+                countDownLatch.countDown();
+            });
+        }
+        try {
+            countDownLatch.await();
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
     }
 
     private void getVmfsDetailFromDme(List<VmfsDataInfo> relists, Map<String, String> stoNameMap,
@@ -2115,14 +2155,16 @@ public class VmfsAccessServiceImpl implements VmfsAccessService {
         throws DmeSqlException {
         // 本地全量查询
         List<String> localWwns = dmeVmwareRalationDao.getAllWwnByType(storeType);
-
+        List<String> storeIds = dmeVmwareRalationDao.getAllStorageIdByType(storeType);
         List<DmeVmwareRelation> newList = new ArrayList<>();
         List<DmeVmwareRelation> upList = new ArrayList<>();
         for (DmeVmwareRelation relation : relationList) {
             String wwn = relation.getVolumeWwn();
-            if (localWwns.contains(wwn)) {
+            String storeId = relation.getStoreId();
+            if (localWwns.contains(wwn) && storeIds.contains(storeId)) {
                 upList.add(relation);
                 localWwns.remove(wwn);
+                storeIds.remove(storeId);
             } else {
                 newList.add(relation);
             }
@@ -2139,7 +2181,7 @@ public class VmfsAccessServiceImpl implements VmfsAccessService {
         }
 
         // 删除
-        if (!localWwns.isEmpty()) {
+        if (!localWwns.isEmpty() && !storeIds.isEmpty()) {
             dmeVmwareRalationDao.deleteByWwn(localWwns);
         }
         return true;
